Relocate functions
